Check migrations on scheduler
