+++ .gitignore
@@ -25,6 +25,7 @@ dist/
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+*.log
 
 src/locales/
 src/react-app.d.ts
