+++ i18n/es.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: i18next-conv\n"
-"POT-Creation-Date: 2021-08-11T07:27:29.533Z\n"
+"POT-Creation-Date: 2021-08-11T07:50:51.125Z\n"
 "PO-Revision-Date: 2018-10-25T09:02:35.143Z\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -199,6 +199,9 @@ msgstr ""
 msgid "Please provide a valid identifier"
 msgstr ""
 
+msgid "Invalid option"
+msgstr ""
+
 msgid "Name"
 msgstr ""
 
@@ -311,9 +314,6 @@ msgid "Skip if all values are missing"
 msgstr ""
 
 msgid "Never skip"
-msgstr ""
-
-msgid "Invalid option"
 msgstr ""
 
 msgid "Saving predictors"
