+++ src/domain/entities/Future.ts
@@ -52,6 +52,15 @@ export class Future<E, D> {
     /* Static methods */
     static noCancel: Cancel = () => {};
 
+    static fromPromise<E, D>(computation: Promise<D>): Future<E, D> {
+        return new Future(
+            fluture.Future((reject, resolve) => {
+                computation.then(data => resolve(data)).catch(error => reject(error));
+                return Future.noCancel;
+            })
+        );
+    }
+
     static fromComputation<E, D>(computation: Computation<E, D>): Future<E, D> {
         return new Future(fluture.Future((reject, resolve) => computation(resolve, reject)));
     }
